# Eliminates Dart->Host copy for mutable responses