--- a/lib/ui/window/platform_message_response_dart.cc
+++ b/lib/ui/window/platform_message_response_dart.cc
@@ -47,6 +47,12 @@ void PostCompletion(Callback&& callback,
 }
 }  // namespace
 
+namespace {
+void MappingFinalizer(void* isolate_callback_data, void* peer) {
+  delete static_cast<fml::Mapping*>(peer);
+}
+}  // anonymous namespace
+
 PlatformMessageResponseDart::PlatformMessageResponseDart(
     tonic::DartPersistentValue callback,
     fml::RefPtr<fml::TaskRunner> ui_task_runner,
@@ -64,9 +70,24 @@ PlatformMessageResponseDart::~PlatformMessageResponseDart() {
 
 void PlatformMessageResponseDart::Complete(std::unique_ptr<fml::Mapping> data) {
   PostCompletion(std::move(callback_), ui_task_runner_, &is_complete_, channel_,
-                 [data = std::move(data)] {
-                   return tonic::DartByteData::Create(data->GetMapping(),
-                                                      data->GetSize());
+                 [data = std::move(data)]() mutable {
+                   void* mapping = data->GetMutableMapping();
+                   Dart_Handle byte_buffer;
+                   if (mapping) {
+                     size_t data_size = data->GetSize();
+                     byte_buffer = Dart_NewExternalTypedDataWithFinalizer(
+                         /*type=*/Dart_TypedData_kByteData,
+                         /*data=*/mapping,
+                         /*length=*/data_size,
+                         /*peer=*/data.release(),
+                         /*external_allocation_size=*/data_size,
+                         /*callback=*/MappingFinalizer);
+
+                   } else {
+                     byte_buffer = tonic::DartByteData::Create(
+                         data->GetMapping(), data->GetSize());
+                   }
+                   return byte_buffer;
                  });
 }