--- a/shell/platform/fuchsia/flutter/file_in_namespace_buffer.h
+++ b/shell/platform/fuchsia/flutter/file_in_namespace_buffer.h
@@ -23,6 +23,9 @@ class FileInNamespaceBuffer final : public fml::Mapping {
   const uint8_t* GetMapping() const override;
 
   // |fml::Mapping|
+  uint8_t* GetMutableMapping() const override { return nullptr; }
+
+  // |fml::Mapping|
   size_t GetSize() const override;
 
   // |fml::Mapping|