This makes `platform_channel_basic_binary_2host_1MB` go from 529.2 µs to 412.4 µs (22% improvement) in my local testing.

The test is in the platform channel benchmarks in the flutter repo.

I found this optimization when exploring asset loading, so it should be beneficial there as well by eliminating a copy.

**WARNING:** Do not merge until https://github.com/flutter/flutter/pull/105982 merges in the framework.

## local testing results
Executed on iOS device.

### before
```
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 56.8 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 684.4 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 49.7 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 529.2 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 26.1 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 39.6 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 22.3 µs
```

### after
```
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 54.4 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 685.3 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 50.1 µs
BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 412.4 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 25.3 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 39.3 µs
BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 22.4 µs
```

## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [x] I read the [Tree Hygiene] wiki page, which explains my responsibilities.
- [x] I read and followed the [Flutter Style Guide] and the [C++, Objective-C, Java style guides].
- [ ] I listed at least one issue that this PR fixes in the description above.
- [ ] I added new tests to check the change I am making or feature I am adding, or Hixie said the PR is test-exempt. See [testing the engine] for instructions on
writing and running engine tests.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] I signed the [CLA].
- [x] All existing and new tests are passing.

If you need help, consider asking for advice on the #hackers-new channel on [Discord].

<!-- Links -->
[Contributor Guide]: https://github.com/flutter/flutter/wiki/Tree-hygiene#overview
[Tree Hygiene]: https://github.com/flutter/flutter/wiki/Tree-hygiene
[Flutter Style Guide]: https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo
[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style
[testing the engine]: https://github.com/flutter/flutter/wiki/Testing-the-engine
[CLA]: https://cla.developers.google.com/
[flutter/tests]: https://github.com/flutter/tests
[breaking change policy]: https://github.com/flutter/flutter/wiki/Tree-hygiene#handling-breaking-changes
[Discord]: https://github.com/flutter/flutter/wiki/Chat