@gaaclarke 2022-06-14
For asset loading the time change for a 1MB payload is: 543.29 µs to 339.95 µs (37%)

Benchmark: https://github.com/flutter/flutter/pull/105982