@chinmaygarde 2022-06-14
I'm totally convinced that not doing the copy is faster and also saves memory. But I'm not sure if this is going to cause segfaults in cases where a PROT_READ only mmaped buffer response is written to from Dart and also doesn't cause threading violations in cases APKAssetMappings.

I was looking into how to verify that this is safe and the guidance to reproduce the contrived use-case would be to duplicate `EmbedderTest.PlatformMessagesCanReceiveResponse`. On the native side, create an mmaped buffer that is write protected. On the Dart side, attempt to write to it. If we segfault, that would mean we need a migration. If not, we need to figure out how we can write to a write protected buffer without copies. Then hunt down those copies. Either way, I feel we'd need a different API and a migration.