@gaaclarke 2022-06-14
I had a chat with @jonahwilliams and @dnfield.  We decided:
1) Having a mutable ByteBuffer that crashes when you write to it is unacceptable
1) Having a *private* channel that elides the copy and we control all the end points isn't good enough since channels are not ever really private and embedders will have to know to follow the convention of not writing to the ByteBuffer that isn't enforced beyond crashing
1) Implementing copy-on-write semantics for the ByteBuffer will probably not work because that would introduce an extra layer of indirection, thus harming performance in other ways.  Dart could probably implement that but immutable data would be sufficient too.
1) We should definitely elide the copy in cases where we are sending over read-write memory (like platform channels)
1) We should land the benchmark for the asset loader
1) We should audit what is actually being sent with "flutter/assets" in `customer: money` to potentially build a case for prioritizing the work on the Dart VM.