@gaaclarke 2022-06-14
Hey, can't the operating system do copy-on-write with mmap?  https://pythonspeed.com/articles/reduce-memory-array-copies/  This would be something to investigate.