@gaaclarke 2022-06-14
> Hey, can't the operating system do copy-on-write with mmap? https://pythonspeed.com/articles/reduce-memory-array-copies/ This would be something to investigate.

Looks like we are already doing copy-on-write (https://github.com/gaaclarke/engine/blob/a236ed5db78c665eb66de540a4fcda57c7862c71/fml/platform/posix/mapping_posix.cc#L75:L75), so this probably isn't a crasher.  I'll have to write the test that Chinmay is suggesting tomorrow.  Potentially we can remove all copies except apk assets which are not threadsafe.