@gaaclarke 2022-06-16
Sorry, pulling this down from review yet again.  While it does functionally what I wanted, I've lost the performance benefit to platform channels on iOS when compared to previous versions.  I need to investigate what happened.  I'm seeing 514µs for `platform_channel_basic_binary_2host_1MB` which is a lot slower than earlier versions.