@gaaclarke 2022-06-16
Okay, I fixed it.  It turns out that `[NSMutableData dataWithBytes:length:]` is 100µs slower than `[NSData dataWithBytes:length:]`.  I'm not sure why, but that unfortunately means we have to use a const_cast to get the speed boost.  That is safe since we are creating those data buffers in memory and essentially passing them off via unique_ptr.  Objective-C doesn't provide tools to assert that statically unfortunately though.