@dnfield 2022-06-16
Dart can't resize bytedata, but the other points stand