@jason-simmons 2022-06-16
I agree with https://github.com/flutter/engine/pull/34018#pullrequestreview-1009730927.  Returning a Dart ByteData backed by a copy-on-write file mapping will produce unexpected behavior.

Is there a specific application scenario that this is targeting?

If the goal is to improve the performance of platform messages sent by host applications, then it might be worthwhile to avoid an extra copy if the message is backed by a malloced buffer or std::vector.

If the issue is related to loading assets, then we probably need a separate Dart API that represents a read-only mapping (as mentioned above).