@gaaclarke 2022-06-17
I've removed the iOS implementation since it will take a bit more work to get it working safely with the BinaryCodec (or custom codecs).