@jonahwilliams 2022-06-17
Is it beneficial to trade off the performance of small messages for larger ones?