@gaaclarke 2022-06-17
> Is it beneficial to trade off the performance of small messages for larger ones?

The previous implementation was already conditionalizing how data was being treated based on its size for performance reasons: https://github.com/gaaclarke/engine/blob/a01f69234e9cc0d21d4a972e4e22bb19737e00fe/third_party/tonic/typed_data/dart_byte_data.cc#L26:L26

I can duplicate that logic here.

(sorry, i edited your comment on accident instead of replying)