@gaaclarke 2022-06-17
@jonahwilliams I duplicated that logic in the latest patch and here are the results for local testing:
```
I/flutter (24598): BasicMessageChannel/StandardMessageCodec/Flutter->Host/Small: 301.4 µs
I/flutter (24598): BasicMessageChannel/StandardMessageCodec/Flutter->Host/Large: 1949.6 µs
I/flutter (24598): BasicMessageChannel/BinaryCodec/Flutter->Host/Large: 124.1 µs
I/flutter (24598): BasicMessageChannel/BinaryCodec/Flutter->Host/1MB: 1055.0 µs
I/flutter (24598): BasicMessageChannel/StandardMessageCodec/Flutter->Host/SmallParallel3: 84.7 µs
I/flutter (24598): BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/Small: 195.6 µs
I/flutter (24598): BasicMessageChannel/StandardMessageCodec/Flutter->Host (background)/SmallParallel3: 127.3 µs
```