@gaaclarke 2022-06-24
One other thought that I didn't communicate well.  The data we are actually sending here in a response is one of 4 types: 
1) Ephemeral data that we can prove at compile time the messenger has unique ownership of (ex MallocMapping)
1) A memory representation of a file (ex mmap)
1) A non-threadsafe memory representation of a file (ex AAssets)
1) Data that we can't statically assert is uniquely owned by the messenger. (ex NSDataMapping)

Only type 4 requires a Dart representation that is immutable.  It also represents a fraction of the actual usage.  Right now it is iOS platform channel usage but I think we can use convention and tests to capture most of iOS's use cases as well by proving that the messenger has unique ownership.