@gaaclarke 2022-06-24
Here is some data supporting the importance of this change. (@zanderso @jonahwilliams @jiahaog )

## Microbenchmarks Summary

Here is a reminder of the microbenchmarks results:

- For iOS asset loading the time change for a 1MB payload is: 543.29 µs to 339.95 µs (37%)
- For Android Platform Channels 1MB round trip time: 1483.9 µs to 1055.0 µs (29%)

## Customer Money

Here is some data that shows the potential gain for customer money.

### Sqflite latency breakdown

In b/237035671 customer money asks why the wall time for loading data from sqflite takes 17ms when the actual execution time of java handler takes 1ms.

The reasons why are:

1) The latency when scheduling on the platform thread (could be removed with https://github.com/tekartik/sqflite/issues/719)
1) The latency when scheduling on sqlflite's thread (probably not huge)
1) The latency when scheduling on the platform thread again (I just removed this in a PR in sqflite: https://github.com/tekartik/sqflite/pull/825)
1) The latency when scheduling the response on the ui thread (This probably contributes the most)
1) The payload size of the request
1) **The payload size of the response (The province of this PR if the response is >1MB)**

I've asked them measure the size of their responses to understand what portion of the 16ms is the response copy.

### Platform channel audit

We do have data from customer money that gives us understanding about how platform channels are being used since the savings from this PR are a function of the size of the response payloads.

Since this improvement is gated on a payload greater than 1000 bytes, here are the platforms channels reported by customer money that are communicating over 1000 bytes on the "downstream" (in the direction of Host->Flutter):

|channel|down_bytes|
|-|-|
|plugins.flutter.io/\<redacted\>#updateConfig|27733|
|plugins.flutter.io/device_info#getAndroidDeviceInfo|5588|
|com.tekartik.sqflite#query|1366|

When we hear back from customer money about the individual sizes we can make sure some of these payloads are over 1000 bytes.

### Asset loading at startup

I also audited asset loading on customer money's prototype app on iOS and these are the assets getting loaded at startup and their size (remember anything over 1000 bytes would receive an improvement):

|asset|size (bytes)|
|-|-|
|asset/ic_us_tab_bar_minus_one_light.flr | **2972**|
|asset/ic_us_tab_bar_zero_light.flr | **2770**|
|asset/ic_us_tab_bar_plus_one_light.flr | **3816**|
|asset/network/home_backsplash_us.svg.vg | **31718**|
|asset/img_logo_\<redacted\>.svg.vg | **4126**|
|asset/network/homescreen_empty_state_business_v2.svg.vg | **2964**|
|asset/network/homescreen_empty_state_people_default_v2.svg.vg | **4412**|
|asset/bg_invite_billboard_us.svg.vg | 154|
|asset/img_invite_billboard_illustration_us.svg.vg | **11077**|
|asset/toolbar/toolbar_footer_light.svg.vg | 69|
|asset/toolbar/toolbar_footer_light.svg.vg | 69|
|asset/my_money/link_to_stored_value_logo.svg.vg | 816|
|asset/entity/container/img_search_icon.svg.vg | **1413**|
|asset/network/remittance_discover_illustration_v3.svg.vg | **5470**|
|asset/network/us_businesses_vertical_gas.svg.vg | **4628**|
|asset/network/us_businesses_vertical_parking.svg.vg | **4178**|
|asset/network/us_businesses_vertical_transit.svg.vg | **2596**|

We cannot easily say how much time this PR would save the app at startup, or generalize for every possible app because it depends on how parallelized the code is and the system is.  We'll get a better idea if we can get them to test the PR.